LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/dynamic-stylesheet-animations-timing-function-expected.txt


PASS The transform property with ease on ::view-transition-group()
PASS The sizing properties with linear on ::view-transition-group()
FAIL Changing the timing function of ::view-transition-group() when animating assert_equals: transform at 50% with linear expected "matrix(1, 0, 0, 1, 200, 0)" but got "matrix(1, 0, 0, 1, 300, 0)"
